+++ src/app/locales/en/components/qos-modal/file-qos-summary.js
@@ -1,0 +1,7 @@
+export default {
+  noRequirements: 'There are no QoS requirements defined for this {{fileType}}',
+  fileType: {
+    file: 'file',
+    dir: 'directory',
+  },
+};