+++ src/app/utils/qos-expression-converters.js
@@ -1,5 +1,15 @@
+/**
+ * Functions and regexp for parsing/visualizing QoS expression
+ * 
+ * @module utils/qos-expression-converters
+ * @author Jakub Liput
+ * @copyright (C) 2020 ACK CYFRONET AGH
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
 import _ from 'lodash';
 
+const nameRe = /^[a-zA-Z0-9]+$/;
 const pairRe = /(.+?)=(.+)/;
 const operatorRe = /(\||&|-)/;
 
@@ -11,12 +21,18 @@ function isOperator(item) {
   return operatorRe.test(item);
 }
 
+function isVariable(item) {
+  return nameRe.test(item);
+}
+
 function expandInfix(obj) {
   if (obj.type === 'group') {
     const parentheses = obj.parentheses;
     return `${parentheses ? '(' : ''}${expandInfix(obj.a)}${operatorChar[obj.operator]}${expandInfix(obj.b)}${parentheses ? ')' : ''}`;
   } else if (obj.type === 'pair') {
     return `${obj.key}=${obj.value}`;
+  } else if (obj.type === 'variable') {
+    return obj.name;
   } else {
     throw new Error(`unrecognized qos object type: ${obj.type}`);
   }
@@ -57,6 +73,11 @@ export function qosRpnToObject(rpnData) {
       } else {
         throw new Error('not enough symbols for operator');
       }
+    } else if (isVariable(item)) {
+      stack.push({
+        type: 'variable',
+        name: item,
+      });
     } else {
       throw new Error(`unrecognized expression element: ${item}`);
     }