+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-6324 Added support for multiple files in QoS modal
 * VFS-5980 Unified Oneprovider GUI Service Pack 2
 * VFS-6232 Added readonly views (eg. metadata) for file browser in preview mode
 * VFS-6270 Fixed not showing shares after adding to new directory